backend/requirements.txt
# Trading Assistant Backend API - Requirements

# Core FastAPI
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5

# Crypto & Market Data
ccxt==4.4.51
pandas==2.2.3
ta==0.11.0

# Web Scraping & HTTP
requests==2.32.5
cloudscraper==1.2.71
feedparser==6.0.11
lxml==5.3.0

# AI & ML
google-generativeai==0.8.3

# Environment & Config
python-dotenv==1.0.1

# CORS
python-multipart==0.0.20

# Optional: For production
gunicorn==23.0.0
